requirements.txt
Django>=5.1,<7.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0
django-crispy-forms>=2.0,<3.0
crispy-bootstrap5>=2024.2,<2026.0
django-filter>=24.0,<25.0
djangorestframework>=3.15,<4.0
python-dotenv>=1.0,<2.0
whitenoise>=6.5,<7.0
gunicorn>=21.2,<23.0
django-allauth[mfa,socialaccount]>=65.0,<66.0
docusign-esign>=6.0,<7.0
anthropic>=0.40,<1.0
